# src/controllers/bookController.ts
import {
  Get,
  Post,
  Route,
  SuccessResponse,
  Body,
  Response,
  Example,
  Delete,
  Path,
  Put,
} from "tsoa";
import { IBook } from "../types/interfaces";
import { Model } from "mongoose";

const BookModel: Model<IBook> = require("../models/book");

@Route("books")
export default class BookController {
  /**
   * Get List of All books
   */
  @Get("/")
  public async getBooks(): Promise<IBook[]> {
    return await BookModel.find();
  }

  /**
   * Get a book details
   * @example bookId "6300e18d3bbd975cf6459994"
   */
  @Response(404, "the requested book in not found")
  @Get("{bookId}")
  public async getBook(bookId: string): Promise<IBook | null> {
    return await BookModel.findById(bookId);
  }
  /**
   * Delete a book
   * @example bookId "6300e18d3bbd975cf6459994"
   */
  @Response(404, "the requested book in not found")
  @SuccessResponse("200", "Deleted")
  @Delete("{bookId}")
  public async deleteBook(bookId: string) {
    return await BookModel.findByIdAndDelete(bookId);
  }

  /**
   * Create a book
   */
  @Response(422, "Validation Failed")
  @SuccessResponse("200", "Created")
  @Example<IBook>({
    title: "Book Tiltle",
    author: "6300e18b3bbd975cf6459983",
    summary: "Book Summary",
    isbn: "Book isbn",
    genre: ["6300e18d3bbd975cf645998e"],
  })
  @Post("create")
  public async createBook(@Body() book: IBook): Promise<IBook> {
    return new BookModel({
      ...book,
    }).save();
  }

  /**
   * Update a book
   */
  @Response(422, "Validation Failed")
  @SuccessResponse("200", "updated")
  @Put("update/{bookId}")
  public async updateBook(
    @Path() bookId: string,
    @Body() book: Partial<IBook>
  ): Promise<IBook | null> {
    let bookDocument = await BookModel.findById(bookId);
    if (bookDocument != null) {
      bookDocument.title = book.title ?? bookDocument.title;
      bookDocument.author = book.author ?? bookDocument.author;
      bookDocument.summary = book.summary ?? bookDocument.summary;
      bookDocument.isbn = book.isbn ?? bookDocument.isbn;
      bookDocument.genre = book.genre ?? bookDocument.genre;
      return await bookDocument.save();
    }
    return null;
  }
}